ADD breadcrumb to header via wrap middleware
ADD css for breadcrumb to header.less
ADD dynamic html for breadcrumb to header.inc.html
ADD data for breadcrumb to templateJSON in wrap middleware

I see some potential problems with this, but these exist before this:
header.inc.html isn't a ejs template right now, it's used and embedded clientside when it doesn't exist yet, as a static html file. With this change that would render improperly, showing the newly added ejs template. But header.inc.html IS already used as a ejs template in wrap.js middleware. So it would seem to be meant as a dynamic template.