ADD dynamic navigation items for header.inc.html
FIX a typo: "heagerFooter" > "headerFooter"
ADD a property to templateJSON object in wrap.js middleware
ADD (placeholder) header navigation items in options.js, it's added into the modulesOptions.headerFooter object. (which was unused until now).
ADD an options property to headerFooter.js, seems to me that it there's a file specific for finding the template files and rendering the header & footer, this is the place to add the data about which navigation items the header contains.